Disable gzip encoding on responses
Gzipped responses are vulnerable to the CRIME and BEAST attacks which could
allow an attacker to partially read a response.

There also may be an issue with zlib allocating memory to unzip responses and
then not returning it to the heap; this could fix the memory problems we've
been seeing on the Autopilot integration.